Only rotate key if it is older than 7 days.

+++ handler_test.go
@@ -1,7 +1,11 @@
 package handler_test
 
 import (
+	"fmt"
 	"testing"
+	"time"
+
+	"github.com/aws/aws-sdk-go/service/secretsmanager"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
@@ -29,13 +33,15 @@ func TestHandler(t *testing.T) {
 	}
 
 	tests := []struct {
-		description string
-		tokenPath   string
-		keyPath     string
-		keyTitle    string
-		team        handler.Team
-		githubKeys  []*github.Key
-		createdKey  *ec2.CreateKeyPairOutput
+		description  string
+		tokenPath    string
+		keyPath      string
+		keyTitle     string
+		team         handler.Team
+		githubKeys   []*github.Key
+		lastUpdated  string
+		shouldRotate bool
+		createdKey   *ec2.CreateKeyPairOutput
 	}{
 
 		{
@@ -50,6 +56,26 @@ func TestHandler(t *testing.T) {
 					Title: github.String("concourse-test-team-deploy-key"),
 				},
 			},
+			lastUpdated:  "2018-01-01T01:01:01Z",
+			shouldRotate: true,
+			createdKey: &ec2.CreateKeyPairOutput{
+				KeyMaterial: aws.String(keyMaterial),
+			},
+		},
+		{
+			description: "does not rotate keys if they have recently been updated",
+			tokenPath:   "/concourse/{{.Team}}/{{.Owner}}",
+			keyPath:     "/concourse/{{.Team}}/{{.Repository}}",
+			keyTitle:    "concourse-{{.Team}}-deploy-key",
+			team:        team,
+			githubKeys: []*github.Key{
+				{
+					ID:    github.Int64(1),
+					Title: github.String("concourse-test-team-deploy-key"),
+				},
+			},
+			lastUpdated:  time.Now().Format(time.RFC3339),
+			shouldRotate: false,
 			createdKey: &ec2.CreateKeyPairOutput{
 				KeyMaterial: aws.String(keyMaterial),
 			},
@@ -63,17 +89,23 @@ func TestHandler(t *testing.T) {
 
 			repos := mocks.NewMockRepoClient(ctrl)
 			repos.EXPECT().ListKeys(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(1).Return(tc.githubKeys, nil, nil)
-			repos.EXPECT().CreateKey(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(1).Return(nil, nil, nil)
-			repos.EXPECT().DeleteKey(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(1).Return(nil, nil)
+			if tc.shouldRotate {
+				repos.EXPECT().CreateKey(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(1).Return(nil, nil, nil)
+				repos.EXPECT().DeleteKey(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(1).Return(nil, nil)
+			}
 
 			apps := mocks.NewMockAppsClient(ctrl)
 			apps.EXPECT().CreateInstallationToken(gomock.Any(), gomock.Any()).Times(1).Return(&github.InstallationToken{Token: github.String("token")}, nil, nil)
 
 			ec2 := mocks.NewMockEC2Client(ctrl)
-			ec2.EXPECT().CreateKeyPair(gomock.Any()).Times(1).Return(tc.createdKey, nil)
-			ec2.EXPECT().DeleteKeyPair(gomock.Any()).Times(1)
+			if tc.shouldRotate {
+				ec2.EXPECT().CreateKeyPair(gomock.Any()).Times(1).Return(tc.createdKey, nil)
+				ec2.EXPECT().DeleteKeyPair(gomock.Any()).Times(1)
+			}
 
 			secrets := mocks.NewMockSecretsClient(ctrl)
+			description := &secretsmanager.DescribeSecretOutput{Description: aws.String(fmt.Sprintf("Github credentials for Concourse. Last updated: %s", tc.lastUpdated))}
+			secrets.EXPECT().DescribeSecret(gomock.Any()).MinTimes(1).Return(description, nil)
 			secrets.EXPECT().CreateSecret(gomock.Any()).MinTimes(1).Return(nil, nil)
 			secrets.EXPECT().UpdateSecret(gomock.Any()).MinTimes(1).Return(nil, nil)
